Add check in the cmp_cfg_icu_buffers function if the compressed data can fit in a compression entity fix some typos; disable 3 times smaller warning if cfg->icu_output_buf = NULL

+++ include/cmp_icu.h
@@ -28,10 +28,10 @@
 struct cmp_cfg cmp_cfg_icu_create(enum cmp_data_type data_type, enum cmp_mode cmp_mode,
 				  uint32_t model_value, uint32_t lossy_par);
 
-size_t cmp_cfg_icu_buffers(struct cmp_cfg *cfg, void *data_to_compress,
-			   uint32_t data_samples, void *model_of_data,
-			   void *updated_model, uint32_t *compressed_data,
-			   uint32_t compressed_data_len_samples);
+uint32_t cmp_cfg_icu_buffers(struct cmp_cfg *cfg, void *data_to_compress,
+			     uint32_t data_samples, void *model_of_data,
+			     void *updated_model, uint32_t *compressed_data,
+			     uint32_t compressed_data_len_samples);
 
 int cmp_cfg_icu_imagette(struct cmp_cfg *cfg, uint32_t cmp_par,
 			 uint32_t spillover_par);
